feat: practitioner health record type
https://new-hippo.atlassian.net/browse/HC-149

+++ sandwich/patients/views/patient/health_records.py
@@ -14,6 +14,8 @@ from django.urls import reverse
 
 from sandwich.core.models import Immunization
 from sandwich.core.models import Patient
+from sandwich.core.models import Practitioner
+from sandwich.core.models.health_record import HealthRecord
 from sandwich.core.service.permissions_service import ObjPerm
 from sandwich.core.service.permissions_service import authorize_objects
 from sandwich.core.util.http import AuthenticatedHttpRequest
@@ -28,11 +30,13 @@ logger = logging.getLogger(__name__)
 def health_records(request: AuthenticatedHttpRequest, patient: Patient):
     documents = patient.document_set.all()
     immunizations = patient.immunization_set.all()
+    practitioners = patient.practitioner_set.all()
 
     context = {
         "patient": patient,
         "documents": documents,
         "immunizations": immunizations,
+        "practitioners": practitioners,
     } | _patient_context(request, patient=patient)
     return render(request, "patient/health_records.html", context)
 
@@ -48,9 +52,31 @@ class ImmunizationForm(forms.ModelForm[Immunization]):
         model = Immunization
         fields = ("name", "date")
 
-    def save(self, *, patient: Patient, commit: bool = True) -> Immunization:  # type: ignore[override]
+    # NOTE: patient is marked as optional here to prevent mypy from complaining that the signature is incompatible
+    #       with the base class, but a database constraint will prevent the form from being submitted without a patient
+    def save(self, commit: bool = True, patient: Patient | None = None) -> Immunization:  # noqa: FBT001,FBT002
         instance = super().save(commit=False)
-        instance.patient = patient
+        if patient is not None:
+            instance.patient = patient
+        if commit:
+            instance.save()
+        return instance
+
+
+class PractitionerForm(forms.ModelForm[Practitioner]):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.helper = FormHelper()
+        self.helper.add_input(Submit("submit", "Submit"))
+
+    class Meta:
+        model = Practitioner
+        fields = ("name",)
+
+    def save(self, commit: bool = True, patient: Patient | None = None) -> Practitioner:  # noqa: FBT001,FBT002
+        instance = super().save(commit=False)
+        if patient is not None:
+            instance.patient = patient
         if commit:
             instance.save()
         return instance
@@ -59,7 +85,9 @@ class ImmunizationForm(forms.ModelForm[Immunization]):
 def _form_class(record_type: str):
     if record_type == "immunization":
         return ImmunizationForm
-    raise ValueError(f"Unknown form class: {record_type}")
+    if record_type == "practitioner":
+        return PractitionerForm
+    raise Http404(f"Unknown form class: {record_type}")
 
 
 @login_required
@@ -82,7 +110,12 @@ def health_records_add(request: AuthenticatedHttpRequest, patient: Patient, reco
     return render(request, "patient/health_records_add.html", context)
 
 
-def _generic_edit_view(record_type: str, request: AuthenticatedHttpRequest, patient: Patient, instance):
+def _generic_edit_view(record_type: str, request: AuthenticatedHttpRequest, instance: HealthRecord):
+    patient = instance.patient
+    if not request.user.has_perms(["view_patient", "change_patient"], instance.patient):
+        # this is the same error that get_object_or_404 raises
+        raise Http404(f"No {instance._meta.object_name} matches the given query.")  # noqa: SLF001
+
     form_class = _form_class(record_type)
     if request.method == "DELETE":
         instance.delete()
@@ -98,6 +131,7 @@ def _generic_edit_view(record_type: str, request: AuthenticatedHttpRequest, pati
             return HttpResponseRedirect(reverse("patients:health_records", kwargs={"patient_id": patient.id}))
     else:
         form = form_class(instance=instance)
+
     context = {
         "record_type": record_type,
         "record_type_name": form_class.Meta.model._meta.verbose_name,  # noqa: SLF001
@@ -108,8 +142,11 @@ def _generic_edit_view(record_type: str, request: AuthenticatedHttpRequest, pati
 
 @login_required
 def immunization_edit(request: AuthenticatedHttpRequest, immunization_id: UUID):
-    model = Immunization
-    instance = get_object_or_404(model, id=immunization_id)
-    if not request.user.has_perms(["view_patient", "change_patient"], instance.patient):
-        raise Http404(f"No {model._meta.object_name} matches the given query.")  # noqa: SLF001
-    return _generic_edit_view("immunization", request, instance.patient, instance)
+    instance = get_object_or_404(Immunization, id=immunization_id)
+    return _generic_edit_view("immunization", request, instance)
+
+
+@login_required
+def practitioner_edit(request: AuthenticatedHttpRequest, practitioner_id: UUID):
+    instance = get_object_or_404(Practitioner, id=practitioner_id)
+    return _generic_edit_view("practitioner", request, instance)
